+++ flopy/modflow/mfwel.py
@@ -247,7 +247,7 @@ class ModflowWel(Package):
         dtype = ModflowWel.get_default_dtype(structured=structured)
         if aux_names is not None:
             dtype = Package.add_to_dtype(dtype, aux_names, np.float32)
-        d = np.zeros((ncells, len(dtype)), dtype=dtype)
+        d = np.zeros((ncells, len(dtype)))
 
         d[:, :] = -1.0E+10
         return np.core.records.fromarrays(d.transpose(), dtype=dtype)
